pyproject.toml
[tool.poetry]
name = "atlassian-oauth-dance"
version = "0.1.0"
description = ""
authors = ["Dennis Stengele <dennis@stengele.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
pycryptodome = "^3.15.0"
requests = "^2.28.1"
requests-oauthlib = "^1.3.1"
oauthlib = {extras = ["signedtoken"], version = "^3.2.0"}
click = "^8.1.3"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
dance = 'atlassian_oauth_dance.cli:cli'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
